# docs/swagger.yaml
consumes:
- application/json
definitions:
  collection.CatalogInfo:
    properties:
      deprecated:
        type: boolean
      name:
        type: string
    type: object
  collection.CollectionInfo:
    properties:
      catalogName:
        type: string
      errorStr:
        type: string
      name:
        type: string
      state:
        type: string
    type: object
  web.CollectionEditPrepinfoResponse:
    properties:
      jsonInput:
        type: string
      schemaJson:
        type: string
    type: object
  web.CollectionEditRequest:
    properties:
      jsonInput:
        type: string
      jsonOutput:
        type: string
      schemaJson:
        type: string
    type: object
  web.CollectionNewRequest:
    properties:
      catalogName:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: zipizap123@gmail.com
    name: Paulo Aleixo Campos
    url: https://github.com/zipizap/vendingMaxine
  description: A focused internal-developer-portal, to support teams getting the best
    of the platform. Simple ;)
  license:
    name: GPL v3
    url: https://github.com/zipizap/vendingMaxine/raw/main/LICENSE
  title: VendingMaxine
  version: "1.0"
paths:
  /api/v1/catalogs:
    get:
      description: Get all catalog overview info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/collection.CatalogInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: All Catalog overview info
      tags:
      - catalog
      - catalogs
  /api/v1/collections:
    get:
      description: Get all collections overview info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/collection.CollectionInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: All Collections overview info
      tags:
      - collection
      - collections
    post:
      consumes:
      - application/json
      description: Collection - create new
      parameters:
      - description: create new Collection
        in: body
        name: newcollection
        required: true
        schema:
          $ref: '#/definitions/web.CollectionNewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Collection - create new
      tags:
      - collection
  /api/v1/collections/{collection-name}/replayable:
    get:
      description: Collection - download .tgz file collectionReplayableTgz
      parameters:
      - description: collection name
        in: path
        name: collectionname
        required: true
        type: string
      produces:
      - application/x-gzip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Collection - download .tgz file collectionReplayableTgz
      tags:
      - collection
  /api/v1/collections/{collectionname}:
    get:
      description: Collection - prepare Collection edit by requesting existing Collection
        info
      parameters:
      - description: prepare Collection edit by requesting existing Collection info
        in: path
        name: collectionname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.CollectionEditPrepinfoResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Collection - prepare Collection edit by requesting existing Collection
        info
      tags:
      - collection
    put:
      consumes:
      - application/json
      description: Collection - initiate Collection edit by putting new Collection
        info
      parameters:
      - description: initiate Collection edit by putting new Collection info
        in: path
        name: collectionname
        required: true
        type: string
      - description: initiate Collection edit by putting new Collection info
        in: body
        name: reqbody
        required: true
        schema:
          $ref: '#/definitions/web.CollectionEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Collection - initiate Collection edit by putting new Collection info
      tags:
      - collection
produces:
- application/json
schemes:
- http
swagger: "2.0"